requirements.txt
fastapi==0.121.0
uvicorn
requests
python-dotenv
pymongo==4.15.3
pandas==2.3.3
httpx==0.28.1
PyJWT==2.10.1
pwdlib[argon2]==0.3.0